4 recording the bare minimum (#15)
* [4] Recording bare minimum

+++ protocol-mongo/src/main/java/org/kendar/mongo/fsm/OpCompressed.java
@@ -62,7 +62,7 @@ public class OpCompressed extends MongoState {
 
 
     private byte[] decompressMessage(byte[] inputBuffer, byte compressorId) {
-        var compressor = compressionHandlers.get(compressorId);
+        var compressor = compressionHandlers.get((int)compressorId);
         if (compressor == null) {
             log.error("Unknow compression " + compressorId);
         }
